Rename component and move validation in the form
