+++ client/auth/src/components/AppLogin.vue
@@ -1,5 +1,5 @@
 <template>
-  <LogistrationWrapper title="Login" :error-messages="errorMessages">
+  <LogistrationForm title="Login" :error-messages="errorMessages">
     <div class="px-8 mt-5">
       <p>Please login</p>
       <div class="p-float-label mt-5 flex">
@@ -41,7 +41,7 @@
         </RouterLink>
       </div>
     </template>
-  </LogistrationWrapper>
+  </LogistrationForm>
 </template>
 
 <script setup>
@@ -50,7 +50,7 @@ import { email, required } from '@vuelidate/validators';
 import authApi from '@/auth/src/api/auth';
 import Button from 'primevue/Button';
 import InputText from 'primevue/inputtext';
-import LogistrationWrapper from './common/LogistrationWrapper.vue';
+import LogistrationForm from './common/LogistrationForm.vue';
 import PasswordInput from './common/PasswordInput.vue';
 import { routes } from '@/shared/utils/navigation';
 import { useVuelidate } from '@vuelidate/core';
